fix. member-service delete 메소드 및 오타 수정, id 초기화 변경

+++ member-service/src/main/kotlin/org/jeongmo/migration/member/infrastructure/adapter/out/jpa/repository/MemberJpaRepository.kt
@@ -31,7 +31,8 @@ class MemberJpaRepository(
 
     override fun delete(id: Long): Boolean {
         var exist = jpaRepository.existsById(id)
+        if (!exist) return false
         jpaRepository.deleteById(id)
-        return exist
+        return true
     }
 }
